\name{fwsim-package}
\alias{fwsim-package}
\docType{package}
\title{
Haplotype tools
}
\description{
Tools for analysing haplotypes, e.g. population simulation.
}
%\details{
%}
\author{
Mikkel Meyer Andersen

Maintainer: Mikkel Meyer Andersen <mikl@math.aau.dk>
}

\keyword{ package }
%\seealso{
%}
%\examples{
%}
